[chain_relay] cleanup
