+++ enclave/chain_relay/src/state.rs
@@ -1,6 +1,7 @@
 use codec::{Decode, Encode};
 use sp_finality_grandpa::{AuthorityList, SetId};
-use sp_runtime::traits::Block as BlockT;
+use sp_runtime::traits::{Block as BlockT, Header as HeaderT};
+use std::fmt;
 use std::vec::Vec;
 
 #[derive(Encode, Decode, Clone, PartialEq)]
@@ -19,5 +20,12 @@ impl<Block: BlockT> RelayState<Block> {
             current_validator_set_id: 0,
             headers: vec![block_header],
         }
+    }
+}
+
+impl<Block: BlockT> fmt::Debug for RelayState<Block> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "RelayInfo {{ last_finalized_block_header_number: {:?}, current_validator_set: {:?}, current_validator_set_id: {} }}",
+               self.last_finalized_block_header.number(), self.current_validator_set, self.current_validator_set_id)
     }
 }
